// _utilities/datatype/Auth/types/screenNames.tsx
export enum ScreenNames {
  email_screen = "email-screen",
  code_screen = "code-screen",
  new_password_screen = "new-password-screen",
  confirm_password_screen = "confirm-password-screen",
  success_screen = "success-screen",
  firstname_screen = "firstname-screen",
  lastname_screen = "lastname-screen",
  dob_screen = "dob-screen",
  image_screen = "image-screen"
}